refactor: significantly improve SVG implementation
