+++ package.json
@@ -39,6 +39,7 @@
     "@fuelrats/eslint-config-react": "^2.1.0",
     "@fuelrats/stylelint-config": "^1.0.1",
     "@rollup/plugin-node-resolve": "^7.1.3",
+    "@svgr/webpack": "^5.4.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "8.1.0",
     "eslint": "^6.8.0",
