--- src/components/svg/BrandSvg.js
@@ -1,30 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const path = 'm289.06069 312.31338c6.28153-5.62032 13.22428-6.83255 18.07318-6.83255 6.28153 0 12.89367-3.08566 17.85277-8.26517 7.49376-7.82436 5.73053-25.23633-2.86526-29.20361-1.87344-0.88162-3.41627-2.20405-3.41627-2.86526 0-0.77142 2.42445-5.62032 5.39992-10.68963 2.97546-5.06931 5.84072-10.79983 6.39173-12.56307 0.99182-3.19586 0.99182-3.19586-1.76324 0.22043-8.37537 10.68962-25.45673 38.90141-42.9789 70.97029-4.07749 7.38356-4.18769 7.82437-0.99182 4.18769 1.98364-2.31425 3.63668-4.18769 4.29789-4.95911m-43.97073 120.89192c-5.39991-2.53462-8.26517-5.9509-10.35901-12.56303-0.99182-3.19589-5.06931-9.03663-9.4774-13.88549-4.29789-4.62853-10.13862-11.46108-13.11408-15.31813-4.8489-6.28153-6.17133-7.16315-13.00388-8.59578-4.18768-0.88165-15.97933-1.54284-26.33835-1.54284-28.10159 0-26.55876-2.20407 2.31425-3.41628 13.66509-0.55101 25.56694-1.43263 26.44856-2.09385 3.19586-1.98363-0.99183-6.72233-7.16316-8.26517-9.9182-2.53465-48.04821-15.42832-47.49719-15.97933 0.33059-0.3306 12.12225 1.87344 26.22815 4.9591 14.10589 3.08567 26.33835 5.51012 27.10977 5.51012 0.77141 0 1.43263-0.88162 1.43263-2.09385 0-1.21222-4.29789-4.18769-10.02841-7.05295-14.1059-7.05295-51.57469-29.42402-53.66853-31.95867-0.99182-1.21223 9.25699 3.52647 22.59147 10.46922 21.93026 11.35084 35.26474 16.86096 40.66466 16.86096 2.75506 0-1.21223-6.28154-11.35084-18.40379-36.58717-43.41972-41.54627-54.66035-40.11364-91.79854 1.10202-28.873 3.63667-45.84416 9.808-65.68058l5.17951-16.64055-3.63667-8.04477c-1.98365-4.40809-4.6285-9.14679-5.95093-10.46922-3.08567-3.41627-8.26517-3.08566-10.13861 0.55101-3.52648 6.61214-0.3306 35.04434 5.0693 45.51356 1.10203 2.20405 1.87344 5.62032 1.65304 7.71416-0.44081 3.41627-0.77142 3.08567-4.29789-3.52647-10.46922-20.05683-15.64873-46.94619-12.01206-62.26431 2.42445-10.24882 9.03659-16.64055 17.08136-16.64055 6.50194 0 8.26518 1.54283 20.49763 17.85277 6.39174 8.48558 9.80801 9.5876 20.82824 6.50194 44.96255-12.56307 83.75377-13.00387 122.43478-1.43263 15.75893 4.7387 18.07318 4.40809 24.46492-2.53465 2.97546-3.19587 5.39991-6.28154 5.39991-6.94275 0-0.55101 2.53466-3.85708 5.73052-7.27335 4.6285-5.06931 6.61214-6.17133 10.46922-6.17133 13.00388 0 19.06499 10.35901 18.7344 32.39948-0.11036 12.34266-0.7714 15.31812-5.62032 28.98321-2.97546 8.37538-6.72234 17.63237-8.37538 20.71804l-2.86526 5.51011-0.11021-5.62032c0-3.08566 0.88162-5.39991 1.43263-7.05295 1.21223-3.85708 6.83255-22.81187 6.83255-32.17907 0-11.90185-0.99183-14.98752-5.06931-15.97934-4.40809-1.10202-8.92639 3.19587-12.89367 12.23246-3.41627 7.93457-3.30607 8.81619 2.75506 27.88118 5.95092 18.8446 10.13861 53.44813 9.25699 75.8192-0.99182 26.66897-6.28153 37.24839-37.46879 75.04778-11.57124 13.9957-16.30994 20.60784-15.42832 21.48946 1.87344 1.87344 16.75075-3.74688 38.35041-14.65691 13.44468-6.83255 24.79552-12.34266 25.34653-12.34266 2.75505 0-3.74688 4.18769-29.31382 18.62419-37.02798 20.93844-35.26474 19.83642-33.61171 21.48945 0.99182 0.99182 9.69781-0.33059 27.33018-4.18768 14.2161-3.08567 26.11795-5.39992 26.33835-5.17951 0.44081 0.44081-27.55058 10.13861-42.09728 14.5467-6.17133 1.87344-10.79983 4.18769-12.12226 5.95093l-2.09384 2.97546 22.37107 1.54283c12.34266 0.88162 24.79552 1.65303 27.77099 1.76324 2.97545 0.11036 4.9591 0.55102 4.40809 1.21221-0.55102 0.55102-12.34266 1.43263-26.00775 1.87344-13.77529 0.44084-26.66896 1.54284-28.76281 2.31428-2.09384 0.7714-6.28153 5.06929-9.4774 9.47739s-8.59578 10.91002-12.12225 14.54668c-3.52648 3.6367-7.60396 9.69782-9.14679 13.66511-4.07749 10.13862-7.93457 12.56307-20.93845 13.11409-7.05294 0.33071-11.90185 0-14.65691-1.32246m-22.37106-132.79378c-12.12225-21.93026-25.89754-44.52174-33.83211-55.43177-6.72234-9.257-5.62032-4.5183 1.87344 8.48558 2.97546 5.06931 5.28971 9.91821 5.28971 10.68962 0 0.77142-1.54283 2.09385-3.41627 2.75506-5.06931 1.98364-8.04477 8.81619-7.38356 17.19157 0.99182 12.34266 8.48558 18.84459 23.8037 20.71803 10.24882 1.21222 14.65691 3.41627 19.39561 9.5876 1.98365 2.64486 3.85708 4.8489 4.07749 4.8489 0.44081 0.11022-4.07749-8.37537-9.80801-18.84459m-147.56091-189.6582c2.75506 0 4.959107 2.20405 4.959107 4.95911s-2.204047 4.9591-4.959107 4.9591c-2.755057 0-4.959103-2.20404-4.959103-4.9591s2.204046-4.95911 4.959103-4.95911m365.6513 1.65304c2.75505 0 4.95912 2.20404 4.95912 4.9591s-2.20407 4.95911-4.95912 4.95911-4.95908-2.20405-4.95908-4.95911 2.20403-4.9591 4.95908-4.9591m-363.66765-22.371075 22.371069-19.946619s4.187685 4.7387 4.187685 4.7387l-8.265168 7.383556 12.232458 13.77529-5.95093 5.289708-12.232454-13.775288-8.265173 7.383556s-4.077487-4.848903-4.077487-4.848903m25.787344-23.473092 6.50193-4.518297s5.06931 7.273353 5.06931 7.273353l13.9957-9.697802s-5.06931-7.273353-5.06931-7.273353l6.50194-4.518297s14.10589 20.387429 14.10589 20.387429l-6.50193 4.518297s-5.06931-7.383557-5.06931-7.383557l-13.9957 9.697803s5.06931 7.383556 5.06931 7.383556l-6.50194 4.518297s-14.10589-20.387429-14.10589-20.387429m33.61171-21.820059 24.57512-12.232459 2.53465 5.179509-17.96298 8.926386s1.54283 3.195871 1.54283 3.195871l16.97116-8.375377 2.42445 4.8489-16.97116 8.375376s1.76324 3.526477 1.76324 3.526477l18.07318-8.92639s2.64485 5.289713 2.64485 5.289713l-24.68531 12.232456-10.91003-22.040462m39.45243-18.624194s25.12613-7.273353 25.12613-7.273353l1.65303 5.620318s-17.96298 5.179509-17.96298 5.179509l1.32243 4.518296 17.08136-4.959105s1.54283 5.510116 1.54283 5.510116-17.08136 4.959103-17.08136 4.959103 2.42445 8.265177 2.42445 8.265177l-7.27335 2.093843-6.83254-23.913904m31.07705-8.265174 7.49376-0.881619 1.21223 11.350839c0.33059 2.644856 0.55101 4.187689 0.77141 4.848903 0.3306 1.102023 1.21223 1.763237 2.64486 1.763237 1.32243 0 3.19587-0.110204 5.84072-0.330607 2.64486-0.330606 4.5183-0.551011 5.62032-0.881618 1.54283-0.44081 2.31425-1.43263 2.31425-2.975463 0-1.322428-0.11021-2.755058-0.22042-4.408093l-1.21223-11.350839s7.49376-0.881618 7.49376-0.881618l1.21223 11.350839c0.33059 3.967283 0.55101 6.501937 0.44081 7.824364-0.11021 2.093845-0.88162 3.746879-2.42446 4.848903-1.21222 0.881617-3.19586 1.54283-6.06112 1.98364-1.32243 0.220422-3.41628 0.44081-6.28154 0.771417-4.18769 0.44081-7.27335 0.661213-9.03659 0.661213-1.32243 0-2.42445-0.110211-3.30607-0.330595s-1.54283-0.661213-2.20404-1.102024c-0.88162-0.771416-1.43263-1.653033-1.76324-2.865259-0.44081-1.322427-0.66121-2.975462-0.99182-4.959104l-0.44081-3.195868-1.10202-11.240636m38.79121-3.636677 27.44038 0.661214-0.11021 5.840723s-20.05682-0.440809-20.05682-0.440809-0.11021 3.526474-0.11021 3.526474 18.9548 0.440809 18.9548 0.440809l-0.11021 5.399914s-18.9548-0.440809-18.9548-0.440809l-0.11021 3.967284 20.16702 0.440809-0.11021 5.950926-27.55058-0.661213 0.55101-24.685322m33.0607 1.102023 7.82436 1.102024-2.64485 18.293585s17.41197 2.534653 17.41197 2.534653l-0.88162 6.281533s-25.34654-3.74688-25.34654-3.74688l3.63668-24.464915m32.50969 30.526045 7.60395-23.583299s17.41197 5.620319 17.41197 5.620319c4.07749 1.322428 6.50194 2.204046 7.16315 2.424451 0.77142 0.330607 1.54283 0.881619 1.98364 1.43263 0.55102 0.661214 0.88162 1.322428 1.10203 2.093844 0.22038 0.771417 0.22038 1.542833 0 2.314249-0.22043 0.881619-0.44081 1.983643-0.88162 3.195866-0.66122 1.87344-1.21223 3.19587-1.65304 3.857083-0.77141 0.99182-1.76323 1.542834-3.08566 1.653033-0.44081 0-0.99182 0-1.87344-0.110211 1.43263 0.55101 2.42445 1.43263 2.86526 2.314247 0.22042 0.44081 0.22042 0.88162 0.22042 1.32243 0 0.440806-0.11021 1.102023-0.33059 1.983639-0.11021 0.44081-0.44081 1.322431-0.88162 2.644857l-0.66122 2.204047-7.05294-2.31425 0.4408-1.212223c0.44081-1.432634 0.77142-2.424454 0.77142-3.085667 0.11021-0.88162-0.22042-1.653037-0.88162-1.983644-0.44081-0.220422-1.43263-0.661213-3.08566-1.212223l-10.13862-3.306069-1.98364 5.950922-7.05294-2.204042m11.02022-9.477402 10.13862 3.306068c1.54283 0.44081 2.42445 0.771417 2.64485 0.771417 0.99183 0.220422 1.76324 0.220422 2.20405-0.330596 0.3306-0.330595 0.66122-0.991819 0.88162-1.873439 0.22038-0.771415 0.3306-1.432629 0.3306-1.873438 0-0.661214-0.55101-1.212226-1.43263-1.542833-0.3306-0.110203-1.32243-0.440809-2.86526-0.991821l-10.13862-3.306069-1.76324 5.840723m44.63194 32.619886v-4.7387l-14.6569-7.493756-3.74688 2.755057-7.71416-3.967283 22.92208-16.08954 10.46922 5.289714 0.33059 27.991388-7.60396-3.74688m0.11021-10.689626v-10.579423l-1.32243-0.661213-8.48557 6.17133 9.808 5.069306m18.6242-7.934566 24.57513 17.301766-3.6367 5.17951-9.03659-6.391737-10.57942 15.097719-6.50193-4.518296 10.57943-15.09772-9.03658-6.391732s3.63666-5.17951 3.63666-5.17951m16.64054 32.840291 5.39992 4.959107c-0.88161 0.881617-1.21221 1.653033-0.99182 2.204047 0.11036 0.330595 0.22035 0.55101 0.44081 0.881616 0.22034 0.220422 0.66122 0.661213 1.21224 1.322427 0.33071 0.330595 1.32242 1.212226 3.08564 2.865263 2.42446 2.204047 3.74688 3.41627 4.18768 3.746876 0.77144 0.551014 1.32246 0.881621 1.65306 0.991821 0.4408 0.11021 1.10203-0.220419 1.65305-0.881617 0.4408-0.44081 0.66119-0.88162 0.55098-1.212227 0-0.55101-0.55098-1.212223-1.54281-2.204047-0.66122-0.661213-2.09385-1.983639-4.18771-4.187686-2.20403-2.204046-3.63666-3.636676-4.29789-4.18769-1.76323-1.653033-2.75505-2.975459-3.19586-4.077483-0.55101-1.32243-0.4408-2.644856 0.33071-3.967286 0.44081-0.771413 0.99182-1.653034 1.98365-2.644853 1.87343-2.093847 3.6367-3.195871 5.1795-3.30607 1.21225-0.110211 2.42445 0.330595 3.85708 1.212223 1.21225 0.771416 3.85709 3.085666 8.0448 6.722343 2.64484 2.42445 4.40807 4.187686 5.39989 5.289709 1.21225 1.542834 1.98365 2.865264 2.09386 3.857084 0.22035 1.542833-0.55102 3.195866-2.20407 4.959103-0.22035 0.220422-0.44081 0.551013-0.88161 0.88162l-5.39989-4.959107c0.33071-0.440806 0.66119-0.771413 0.7714-0.99182 0.33071-0.661213 0.22035-1.322426-0.4408-1.983639-0.55102-0.661217-2.09386-1.983644-4.40811-4.18769-2.31424-2.093843-3.85708-3.30607-4.4081-3.636677-0.66119-0.330595-1.2122-0.110211-1.87343 0.551014-0.55102 0.661213-0.66119 1.322426-0.33071 1.873436 0.33071 0.551014 2.20403 2.424454 5.62031 5.730523 3.08568 2.975464 5.06933 4.959103 5.84073 5.840723 0.7714 0.99182 1.32242 1.763237 1.54284 2.534654 0.33071 0.771416 0.33071 1.542833 0.22035 2.314246-0.22035 1.212224-0.99183 2.534654-2.42446 4.187694-1.65301 1.76323-3.08564 2.86526-4.29789 3.30607-1.21221 0.44081-2.53462 0.33059-3.85708-0.3306-1.32242-0.66121-4.29789-3.08567-9.03659-7.383556-2.86526-2.534653-4.8489-4.518297-5.9509-5.840723-1.43263-1.542833-2.20407-2.975463-2.42445-4.077487-0.22035-1.212222 0.11036-2.424449 0.88161-3.526473 0.33071-0.551013 0.88161-1.212226 1.43263-1.87344 0 0 0.7714-0.771416 0.7714-0.771416m-146.4588-83.202775c-141.38958 0-255.99999 114.61041-255.99999 255.99999s114.61041 256.00001 255.99999 256.00001 256.00001-114.61043 256.00001-256.00001-114.61043-255.99999-256.00001-255.99999m0 7.7141625c137.09169 0 248.28585 111.19414 248.28585 248.28583s-111.19416 248.28584-248.28585 248.28584-248.28583-111.19415-248.28583-248.17564c0-136.98148 111.19414-248.39603 248.28583-248.39603m0 36.807574c-117.14506 0-212.02926 94.994403-212.02926 212.02926 0 117.14507 94.994402 212.02926 212.02926 212.02926 117.14507 0 212.02926-94.99439 212.02926-212.02926 0-117.14506-94.88419-212.02926-212.02926-212.02926m0 7.163154c113.17779 0 204.86611 91.688329 204.86611 204.86611 0 113.17779-91.68832 204.86611-204.86611 204.86611-113.17778 0-204.86611-91.68832-204.86611-204.86611 0-113.06758 91.798533-204.86611 204.86611-204.86611m-215.88635 117.69608s-2.975462 6.06112-2.975462 6.06112c-0.991821-0.4408-1.653034-0.77141-1.983641-0.88161-1.43263-0.77142-2.424451-1.54284-3.085665-2.42446-0.661214-0.88161-0.881619-2.09384-0.551012-3.63667 0.330607-1.65304 1.43263-4.29789 3.195867-7.82436 1.983642-3.96729 3.30607-6.61214 4.187687-7.82437 0.771416-1.21222 1.432633-2.09384 1.983643-2.75506 0.551014-0.66121 1.212227-1.10202 1.87344-1.32243 0.88162-0.33059 1.87344-0.33059 2.975463 0 1.102024 0.3306 3.195867 1.21223 6.061127 2.64486 2.865259 1.43263 4.738699 2.42445 5.620319 3.08567 1.322427 0.99182 2.093844 2.09384 2.314247 3.30607 0.110211 0.99182-0.110211 2.31424-0.661213 3.96728-0.55101 1.76324-1.98364 4.62849-4.077487 8.81618-1.212223 2.53466-2.204047 4.18769-2.86526 5.28971-0.99182 1.54284-1.87344 2.53466-2.644856 3.08567-0.99182 0.66121-2.093844 0.77142-3.306067 0.44081-0.88162-0.22042-1.76324-0.55101-2.75506-1.10202-0.220422-0.11021-0.551009-0.3306-1.102023-0.55102l2.975463-6.06112c0.44081 0.22042 0.88162 0.33059 0.99182 0.4408 0.661213 0.22043 1.102023 0.11022 1.542833-0.22042 0.551014-0.55101 1.653034-2.42445 3.195867-5.39991 0.991824-2.09384 1.653037-3.52648 1.983643-4.40809 0.330596-0.99182 0.110211-1.87344-0.661213-2.53466-0.661217-0.55101-1.76324-1.21222-3.416273-2.09384s-2.86526-1.32243-3.636677-1.43263c-0.99182-0.22042-1.763236 0.22042-2.424449 1.21222-0.551014 0.77142-1.322431 2.20405-2.424454 4.4081-1.102023 2.31425-1.87344 3.85708-2.204047 4.7387-0.330595 0.88162-0.220422 1.43263 0.110212 1.87344 0.220422 0.22042 0.771416 0.55101 1.763236 1.10202m5.62032 27.33017c-1.212227 4.4081-2.204047 7.49376-2.975463 9.1468-0.99182 2.42445-2.644857 3.52647-4.848904 3.52647-1.542831 0-3.85708-0.44081-7.16315-1.43263-3.306077-0.9918-5.620326-1.76322-6.832551-2.53463-1.87344-1.10203-2.644856-2.97547-2.204047-5.51012 0.220405-1.76324 0.991821-4.9591 2.314249-9.5876s2.314249-7.71416 2.975463-9.3672c0.991821-2.42445 2.644856-3.52647 4.848902-3.52647 1.542833 0 3.857081 0.44081 7.16315 1.43263 3.30607 0.99182 5.62032 1.76323 6.942747 2.53465 1.873439 1.10202 2.644856 2.97546 2.204046 5.51012-0.330595 1.87344-1.102023 5.0693-2.42445 9.808m-5.399913-2.09384c0.881617-3.19587 1.43263-5.17951 1.43263-5.95093 0-0.88161-0.330595-1.54283-0.991823-1.98364-0.661214-0.44081-2.093843-0.99182-4.077485-1.54283-1.322428-0.33059-2.314249-0.55101-2.86526-0.55101-1.102023 0-1.983642 0.55101-2.534654 1.87344-0.440809 0.88162-0.991821 2.75506-1.763237 5.51011-0.771416 2.75506-1.212225 4.6285-1.43263 5.62032-0.220403 1.32243 0.220403 2.31425 1.212226 2.86526 0.551011 0.3306 1.43263 0.66122 2.644855 0.99182 1.542833 0.44081 2.644856 0.66122 3.30607 0.77142 1.102023 0.11021 1.873439 0 2.424451-0.44081 0.440811-0.3306 0.991821-1.32243 1.432631-2.75506 0.110211-0.55101 0.551009-1.98364 1.212226-4.40809m0.661214 18.73439s-1.102024 6.94275-1.102024 6.94275-16.199742-2.64485-16.199742-2.64485l-2.534653 15.42832s-5.510116-0.88162-5.510116-0.88162l3.636676-22.48127s21.709859 3.63667 21.709859 3.63667m-3.636678 23.2527-0.220403 7.05294-16.420146-0.44081-0.440809 15.64873-5.620319-0.11021 0.661214-22.70168s22.040465 0.55102 22.040465 0.55102m0.330606 39.56263c0.330607 4.6285 0.44081 7.82437 0.330607 9.5876-0.110211 2.64486-1.21222 4.29789-3.306064 4.9591-1.432632 0.55102-3.857084 0.88162-7.273355 1.10203-3.416272 0.22042-5.840723 0.22042-7.383556-0.11021-2.093848-0.44079-3.416276-1.87342-3.967288-4.40808-0.330607-1.76323-0.771416-4.9591-1.102023-9.808-0.330607-4.7387-0.44081-8.04477-0.330607-9.80801 0.110203-2.64486 1.212225-4.29789 3.30607-4.95911 1.43263-0.55101 3.857081-0.88161 7.273353-1.10202 3.416272-0.22038 5.950925-0.22038 7.383555 0.11021 2.093844 0.44081 3.416272 1.87344 3.967284 4.40809 0.440809 1.76324 0.771416 5.06931 1.102023 10.02841m-5.730521 0c-0.220404-3.30607-0.440809-5.39992-0.661214-6.06113-0.330607-0.88162-0.881618-1.32243-1.653034-1.54283-0.771417-0.22039-2.204047-0.22039-4.408093 0-1.322428 0.11021-2.314249 0.22042-2.975463 0.44081-1.102023 0.33059-1.653035 1.21222-1.763237 2.53465-0.110204 0.99182 0 2.97546 0.220403 5.73052s0.440809 4.7387 0.661214 5.73052c0.220403 1.32243 0.991821 2.09385 2.093844 2.31425 0.661214 0.11021 1.542832 0.11021 2.86526 0 1.542833-0.11021 2.644856-0.22042 3.30607-0.3306 1.102023-0.22038 1.763237-0.66121 2.093844-1.21222s0.440809-1.54283 0.440809-3.08567c0.110204-0.66121 0-2.09384-0.220403-4.51829m-1.212225 39.89323 1.43263 6.61214c-0.991821 0.22042-1.763237 0.44081-2.093844 0.55101-1.542833 0.33059-2.865261 0.33059-3.967284 0.11021-1.102023-0.33062-1.983641-1.10203-2.755058-2.53466-0.771416-1.54284-1.542832-4.18769-2.424451-8.04477-0.991821-4.29789-1.542832-7.16315-1.653035-8.70598-0.220403-1.43263-0.220403-2.53466-0.110203-3.41628 0.110203-0.88161 0.330606-1.54283 0.661214-2.20404 0.551011-0.88162 1.212225-1.43263 2.314249-1.87344 1.102023-0.44081 3.195867-0.99182 6.391734-1.76324 3.085665-0.66121 5.179509-0.99182 6.281533-1.10202 1.653034-0.11022 2.975465 0.22042 3.857082 1.10202 0.771416 0.66121 1.322426 1.76324 1.983643 3.41627 0.661213 1.76324 1.43263 4.84891 2.42445 9.4774 0.551013 2.75506 0.99182 4.6285 1.102023 5.84072 0.220422 1.76324 0.110211 3.08567-0.220422 4.07749-0.44081 1.10202-1.212227 1.98364-2.31425 2.42445-0.771417 0.3306-1.763237 0.66121-2.86526 0.88162-0.220422 0-0.551013 0.11021-1.212227 0.22042 0 0-1.432631-6.61214-1.432631-6.61214 0.551012-0.11021 0.881619-0.22038 1.102024-0.33059 0.661214-0.22039 0.991821-0.55101 0.991821-1.10203 0.110211-0.77141-0.220422-2.86526-0.991821-6.17132-0.44079-2.31429-0.881599-3.85711-1.212206-4.62853-0.330607-0.99183-1.102023-1.54283-2.093844-1.54283-0.881618 0-2.204046 0.11021-3.967283 0.551-1.87344 0.44082-3.085666 0.77142-3.746879 1.10203-0.881619 0.44081-1.212226 1.21222-1.102024 2.42445 0.110204 0.99182 0.330607 2.53465 0.881619 4.9591 0.551012 2.53466 0.991821 4.18769 1.322428 5.06931 0.330606 0.88162 0.661214 1.32243 1.322428 1.43263 0.330606 0.11021 1.102023-0.11021 2.093844-0.22042m3.416272 35.04434s2.755058-3.19587 2.755058-3.19587l-5.399914-13.44468-4.187688-0.3306-2.86526-7.16315s24.795521 2.64486 24.795521 2.64486l3.857084 9.6978-16.089541 18.8446-2.86526-7.05295m6.391735-7.16315s6.171329-7.05295 6.171329-7.05295l-0.55101-1.21222-9.367199-0.77142s3.74688 9.03659 3.74688 9.03659m-2.093844 17.08136 19.285406-10.68962 7.934566 14.21609c1.87344 3.30607 2.975464 5.28972 3.195867 5.95093 0.330595 0.77142 0.44081 1.43263 0.44081 2.20405 0 0.77141-0.220384 1.43263-0.55101 1.98364-0.330595 0.66121-0.88162 1.10202-1.43263 1.54283-0.661217 0.44081-1.432634 0.99182-2.534653 1.54283-1.542833 0.88162-2.75506 1.43263-3.416273 1.54285-1.102024 0.22034-2.093844-0.11037-2.975464-0.77144-0.330595-0.22035-0.661213-0.6612-1.102023-1.21222 0.551014 1.32243 0.771417 2.31425 0.44081 3.19586-0.110211 0.44081-0.330595 0.77144-0.661213 0.99183-0.330595 0.33071-0.771417 0.66123-1.43263 1.10203-0.330595 0.22035-1.102023 0.55102-2.093847 1.21221l-1.763236 0.99182-3.195867-5.73052 0.99182-0.55101c1.102023-0.6612 1.983643-1.10201 2.424454-1.54282 0.661213-0.55101 0.881616-1.21223 0.661213-1.87344-0.110211-0.44081-0.551014-1.32243-1.32243-2.64486 0 0-4.628497-8.26517-4.628497-8.26517l-4.848903 2.75505-3.41627-5.95092m12.89367 0.3306s4.6285 8.26517 4.6285 8.26517c0.661213 1.21223 1.102023 1.98364 1.322426 2.20405 0.551014 0.77141 1.102023 1.21222 1.653037 1.10202 0.440806 0 0.99182-0.22038 1.763236-0.66121 0.661214-0.3306 1.102024-0.66122 1.322427-0.99182 0.330595-0.44081 0.330595-1.10203 0-1.87344-0.110211-0.22043-0.55101-1.10203-1.322427-2.42446 0 0-4.6285-8.26517-4.6285-8.26517l-4.738699 2.64486m20.828239 14.5467s14.216099 19.83645 14.216099 19.83645l-4.187686 2.97544-10.35902-14.43648-2.534653 1.87344 9.808006 13.77528-3.967283 2.86526-9.808006-13.77528-2.86526 2.09382 10.359016 14.54672-4.29789 3.08564s-14.216099-19.9466-14.216099-19.9466l17.852776-12.89369m31.95867 38.57081c3.30607 3.1959 5.51012 5.51014 6.61214 6.94277 1.65304 1.98365 1.87344 3.96729 0.88162 5.9509-0.77142 1.32246-2.31425 3.1959-4.6285 5.73056-2.42445 2.42442-4.18769 4.07747-5.51011 4.84887-1.873442 1.10203-3.857081 0.88165-5.950928-0.66119-1.43263-1.10203-3.85708-3.30607-7.273353-6.61213-3.41627-3.30611-5.73052-5.62035-6.832543-7.05298-1.653037-1.98365-1.87344-3.96726-0.88162-5.9509 0.771416-1.32242 2.31425-3.1959 4.6285-5.62035 2.42445-2.53463 4.29789-4.07747 5.510116-4.84887 1.873437-1.10203 3.85708-0.88165 5.950923 0.66119 1.542833 1.10204 3.967285 3.30607 7.493755 6.61213m-4.297889 3.85709c-2.424449-2.31424-3.967283-3.74687-4.628496-4.07747-0.771417-0.44081-1.542833-0.44081-2.314246-0.11036-0.771417 0.4408-1.76324 1.32242-3.30607 2.86526-0.991824 0.99182-1.542833 1.76322-1.87344 2.31424-0.551013 0.99182-0.44081 1.98365 0.440806 3.08568 0.661217 0.7714 1.983644 2.20403 3.967287 4.18768 2.093843 1.98365 3.526473 3.30607 4.29789 3.85708 1.102023 0.88162 2.093843 0.88162 3.085663 0.33071 0.551015-0.33071 1.212225-0.99182 2.204045-1.87343 1.10202-1.102 1.87344-1.98365 2.20405-2.53463 0.66121-0.88165 0.88162-1.65305 0.77141-2.31428-0.11021-0.55098-0.66121-1.43263-1.76323-2.53463-0.3306-0.4408-1.32243-1.54284-3.085669-3.19589m25.456739 30.63625 5.51012 3.85708c-0.55102 0.88161-0.99183 1.54284-1.21223 1.87343-0.88162 1.32246-1.87344 2.20407-2.86526 2.75509s-2.20405 0.55102-3.63668 0c-1.54283-0.55102-4.07748-1.98365-7.27335-4.18771-3.63668-2.53463-6.06113-4.18768-7.16315-5.17951-1.10203-0.88161-1.87344-1.65301-2.42445-2.31424-0.55102-0.66123-0.88162-1.32242-0.99182-2.09386-0.22042-0.99178-0.11021-1.98361 0.33059-2.97543 0.44081-1.10204 1.65304-2.97547 3.52648-5.62035 1.76323-2.64484 3.08566-4.29789 3.85708-5.06929 1.10202-1.21221 2.31425-1.87343 3.63668-1.87343 0.99182 0 2.20404 0.3307 3.85708 1.2122 1.65303 0.88165 4.40809 2.53467 8.26517 5.17954 2.31425 1.54281 3.85708 2.75505 4.8489 3.52645 1.32243 1.10204 2.20405 2.20407 2.64486 3.08569 0.44081 1.10199 0.44081 2.20403 0 3.41627-0.3306 0.7714-0.77142 1.65302-1.43263 2.64484-0.11021 0.11036-0.3306 0.55102-0.77142 0.99182l-5.51011-3.85708c0.33059-0.44081 0.44081-0.7714 0.55101-0.99182 0.33059-0.55102 0.33059-1.10204 0-1.54284-0.44081-0.66119-2.09385-1.98365-4.95911-3.85709-1.87344-1.32242-3.30607-2.20403-4.07748-2.53462-0.99182-0.44081-1.87344-0.33071-2.64486 0.3307-0.66121 0.55102-1.43263 1.65306-2.53465 3.08569-1.10203 1.5428-1.76324 2.64483-1.98364 3.41624-0.3306 0.88165 0 1.76326 0.88161 2.53466 0.66122 0.66123 2.09385 1.65305 3.96729 2.97547 2.09384 1.43263 3.63668 2.42445 4.40809 2.86526 0.77142 0.44081 1.43263 0.44081 1.87344 0.22035 0.22042-0.55098 0.77142-1.102 1.32243-1.87344m27.10977 15.86914 5.84072 3.30607c-0.44081 0.99182-0.77141 1.65305-0.99182 1.98364-0.77142 1.43263-1.65303 2.42446-2.53465 2.97547-0.99182 0.66123-2.20405 0.77141-3.63668 0.33071-1.65303-0.4408-4.18769-1.54284-7.71416-3.52645-3.85708-2.09385-6.39173-3.6367-7.71416-4.51831-1.21223-0.7714-2.09385-1.54284-2.64486-2.09386-0.55101-0.66119-0.99182-1.32242-1.21222-1.98361-0.3306-0.88165-0.22043-1.87343 0.11021-2.97547 0.33059-1.10203 1.32242-3.08568 2.97546-5.95094 1.54283-2.75505 2.64485-4.51827 3.30607-5.39989 0.99182-1.32245 2.09384-2.09385 3.41627-2.20406 0.99182-0.11037 2.31425 0.22034 3.85708 0.77143 1.65304 0.6612 4.6285 2.09383 8.70599 4.40807 2.42445 1.32245 4.18768 2.42445 5.0693 3.08568 1.43263 0.99182 2.42445 1.98365 2.86526 2.75505 0.55101 1.10203 0.66121 2.20407 0.3306 3.30606-0.22043 0.88162-0.66122 1.76327-1.21223 2.75505-0.11021 0.22035-0.3306 0.55102-0.66121 1.10204l-5.84073-3.30607c0.22043-0.4408 0.44081-0.7714 0.44081-0.99182 0.22043-0.66119 0.22043-1.10204-0.22042-1.54284-0.55101-0.55102-2.31424-1.76323-5.28971-3.41628-1.98364-1.102-3.41627-1.87344-4.29789-2.20403-0.99182-0.44081-1.87344-0.22035-2.64485 0.55102-0.55101 0.66122-1.32243 1.76322-2.20405 3.30606-0.88162 1.65305-1.43263 2.86526-1.65303 3.52649-0.22043 0.99182 0.11021 1.76323 1.10202 2.42445 0.77141 0.55098 2.20405 1.43263 4.29789 2.53463 2.20405 1.21225 3.85708 2.09386 4.6285 2.42445 0.77141 0.33071 1.43263 0.33071 1.98364 0 0.66121 0 1.10202-0.55101 1.54283-1.43263m28.21179 21.59966-0.44081-4.18767-13.55488-5.17951-3.08566 2.75505-7.16315-2.75505 18.84459-16.42015 9.6978 3.74687 2.86526 24.68534-7.16315-2.64488m-0.88161-9.47739-0.99183-9.36722-1.21222-0.4408-6.94275 6.28153 9.1468 3.52649m12.12225 5.5101 6.28153 1.54284c-0.22042 1.10203-0.22042 1.76323 0.11021 2.20407 0.22043 0.22035 0.44081 0.33071 0.66122 0.44081 0.33059 0.10998 0.77141 0.3307 1.54283 0.4408 0.3306 0.11036 1.54283 0.44081 3.63668 0.88161 2.75505 0.66123 4.40809 0.99183 4.8489 1.10204 0.88162 0.10998 1.43263 0.10998 1.76324 0 0.4408-0.11036 0.77141-0.66123 0.88161-1.43263 0.11021-0.55102 0.11021-0.99182-0.11021-1.21225-0.22042-0.3307-0.99182-0.7714-2.09384-0.99182-0.77142-0.22035-2.53466-0.66119-5.06931-1.43263-2.64486-0.7714-4.40809-1.21221-5.17951-1.43263-2.09384-0.55102-3.41627-1.10203-4.18769-1.76322-0.99182-0.77144-1.43263-1.87344-1.54283-3.19586 0-0.77144 0.11021-1.76326 0.3306-2.86526 0.55101-2.42445 1.43263-4.07751 2.64485-4.84891 0.88162-0.55101 1.98364-0.88165 3.52647-0.77144 1.32244 0.11037 4.4081 0.66123 9.1468 1.76327 3.08566 0.7714 5.28971 1.32242 6.50193 1.76322 1.65304 0.66123 2.75506 1.32242 3.30607 2.09386 0.77142 1.102 0.99183 2.64484 0.44081 4.84887 0 0.22035-0.11021 0.55102-0.22042 1.10203l-6.28153-1.54284c0.11021-0.55101 0.11021-0.88161 0.11021-1.10199 0-0.66123-0.44081-1.10204-1.21223-1.32246-0.66121-0.22035-2.42445-0.66119-5.28971-1.32242-2.75506-0.66119-4.40809-0.99182-4.9591-0.88161-0.66122 0-0.99182 0.4408-1.21223 1.21221-0.22042 0.77144 0 1.32245 0.55101 1.54284 0.44081 0.22034 2.75506 0.88161 6.83255 1.98364 3.74688 0.99183 6.06113 1.65306 7.05295 2.09386 0.99182 0.44081 1.76323 0.88162 2.31424 1.32242 0.55102 0.44081 0.88162 0.99182 1.21223 1.65302 0.33059 0.99182 0.33059 2.42445-0.11021 4.29792-0.44081 2.09382-1.10202 3.63666-1.87344 4.40807-0.77141 0.88161-1.87344 1.32245-3.19587 1.32245-1.32242 0.10999-4.7387-0.55101-10.13861-1.87347-3.30607-0.7714-5.73052-1.43263-7.16315-1.98361-1.76324-0.66123-2.97546-1.32246-3.63668-2.09386-0.66121-0.88161-0.99182-1.87344-0.88162-3.08568 0-0.55098 0.22043-1.32242 0.3306-2.09382l0.33059-0.77144m32.06888-8.26515s6.94275 0.99183 6.94275 0.99183l-2.97546 21.82004s-6.83255-0.99182-6.83255-0.99182l2.86526-21.82005m25.23633 2.31425c4.6285 0.22034 7.82437 0.4408 9.5876 0.7714 2.53466 0.4408 4.07749 1.76326 4.5183 3.85708 0.3306 1.43263 0.44081 3.85708 0.22038 7.38357-0.11021 3.41628-0.44081 5.84073-0.88162 7.27336-0.66121 2.09382-2.31424 3.19586-4.84889 3.41624-1.76324 0.11036-5.06931 0.11036-9.80801-0.11036-4.7387-0.22035-8.04477-0.44081-9.80801-0.77144-2.53465-0.44081-4.07748-1.76322-4.51829-3.85708-0.3306-1.43263-0.44081-3.85709-0.22043-7.38354 0.11022-3.41627 0.44081-5.84073 0.88162-7.27336 0.66122-2.09386 2.31425-3.19585 4.84891-3.41627 1.76323-0.22035 5.0693-0.22035 10.02841 0.11036m-0.77142 5.73052c-3.41627-0.11036-5.39991-0.11036-6.17133 0-0.88162 0.22034-1.43263 0.66119-1.76324 1.43263-0.22038 0.7714-0.44081 2.20403-0.55101 4.29789-0.11021 1.32242 0 2.31424 0.11022 2.97547 0.22042 1.10199 0.99182 1.76322 2.31425 1.98364 0.99182 0.22035 2.86526 0.33071 5.73051 0.44081 2.86526 0.11036 4.7387 0.11036 5.73053 0.11036 1.32242-0.11036 2.20404-0.66119 2.53465-1.76322 0.22039-0.55102 0.3306-1.54284 0.3306-2.86526 0.11021-1.54284 0.11021-2.75505 0-3.41628-0.11021-1.10203-0.44081-1.87343-0.88162-2.20403-0.44081-0.33071-1.43263-0.66123-2.97547-0.77144-0.4408 0-1.98364-0.22035-4.40809-0.22035m16.86096-5.28971 11.68144-1.43263s13.44468 14.87732 13.44468 14.87732l0.66122-0.11037-2.09385-16.30994 6.83255-0.7714 2.64485 21.93025-11.57124 1.43263s-13.44468-14.98753-13.44468-14.98753l-0.77142 0.11037 2.09385 16.30994-6.83255 0.7714-2.64485-21.82004m32.61988-4.4081s23.5833-6.28154 23.5833-6.28154l1.32243 4.95912s-17.19156 4.62849-17.19156 4.62849l0.77141 3.08568s16.30995-4.29789 16.30995-4.29789l1.21222 4.62848-16.30994 4.29789 0.88162 3.41628s17.30176-4.62848 17.30176-4.62848l1.32243 5.06929s-23.6935 6.28153-23.6935 6.28153l-5.51012-21.15885m25.34654-6.06111s11.35084-4.62853 11.35084-4.62853l11.46104 10.79985 0.66121-15.75893 11.24064-4.62852 8.37537 20.38745-6.17133 2.53467-6.06112-15.31813-0.99182 0.44081-1.43263 18.40377-5.17951 2.09385-13.88549-12.12226-1.10203 0.4408 6.39174 15.20791-6.17133 2.53467-8.48558-20.38741m43.86052-20.71805 5.84072-3.9673 9.14681 13.5549s13.00388-8.8162 13.00388-8.8162 3.19585 4.62852 3.19585 4.62852l-18.84461 12.89367-12.34265-18.29359m20.93843-14.21609 5.62035-4.29789 13.33447 17.52215-5.6203 4.29789-13.33452-17.52215m23.25271 10.13862-14.87731-16.30995 11.68146-10.68963c2.53463-2.31425 3.96726-3.63666 4.51827-3.9673 1.10204-0.88161 2.09386-1.43263 2.86526-1.54284 0.88162-0.11036 1.76323 0 2.53467 0.44085 0.7714 0.4408 1.43263 1.10199 2.09382 1.87343 0.88165 0.99183 1.43263 1.76323 1.65305 2.42446 0.44081 1.10199 0.22035 2.20403-0.44081 3.41623 0.88162-0.55098 1.54284-0.88161 2.09386-0.88161 1.32242-0.22034 2.75505 0.55102 4.29789 2.31424 1.21221 1.32246 1.87344 2.53467 2.09382 3.52649 0.11036 0.88162-0.11036 1.76323-0.66119 2.64488-0.44081 0.7714-1.43263 1.76322-2.97547 3.19585l-3.19585 2.97547-11.68147 10.57943m-4.29789-13.9957 6.94273-6.39175c1.54284-1.43263 2.42446-2.31424 2.53467-2.64484 0.22034-0.55102 0.11036-1.10203-0.44081-1.65305-0.55102-0.66119-1.21221-0.7714-1.76323-0.55102-0.22034 0.11036-1.10203 0.88161-2.64487 2.31424l-6.94273 6.39175 2.31424 2.53467m5.5101 5.84073 7.05298-6.39175c1.32242-1.21225 2.09382-1.87344 2.31424-2.09386 0.66119-0.66119 0.99182-1.21221 0.99182-1.65301 0-0.33071-0.22034-0.77144-0.66122-1.21225-0.66123-0.66119-1.21221-0.88161-1.87344-0.55102-0.22035 0.11037-1.32242 1.10204-3.19586 2.86526l-7.05297 6.39175 2.42445 2.64488m5.62035-28.21179s15.97931-18.51402 15.97931-18.51402l3.9673 3.41628-11.57125 13.44468 2.42445 2.09386 11.02023-12.78349 3.63666 3.19589-11.02023 12.78346 2.64488 2.31424 11.68146-13.5549 3.96725 3.41628-15.97931 18.62419-16.75075-14.43647m35.26473-7.71417-18.62419-11.79167 8.70599-13.6651c2.09382-3.19586 3.30607-5.0693 3.74687-5.62031 0.44081-0.66122 1.10204-1.10203 1.76323-1.43263 0.66123-0.3306 1.32242-0.55102 1.98364-0.44081 0.77141 0 1.43263 0.22038 1.98365 0.55101 0.7714 0.33059 1.54284 0.88162 2.53467 1.43263 1.5428 0.99182 2.53462 1.76324 2.97543 2.31425 0.66123 0.88162 0.88165 1.87344 0.66123 2.97546-0.11036 0.3306-0.22035 0.88162-0.55102 1.54283 0.88161-1.10202 1.76323-1.76323 2.64488-1.87344 0.4408-0.11021 0.7714 0 1.2122 0.11021 0.44081 0.11022 0.99183 0.3306 1.65306 0.77142 0.3307 0.22042 0.99182 0.66121 2.09382 1.32243l1.76326 1.10202s-3.52649 5.62031-3.52649 5.62031l-0.99182-0.66119c-1.10204-0.66123-1.98365-1.21224-2.53467-1.32245-0.7714-0.33071-1.43263-0.22035-1.87343 0.3307-0.33071 0.33071-0.88162 1.102-1.65302 2.42446 0 0-5.06933 7.93455-5.06933 7.93455l4.7387 2.97547s-3.63666 5.39992-3.63666 5.39992m-5.73052-11.57125s5.06929-7.93458 5.06929-7.93458c0.77144-1.21222 1.21225-1.87344 1.32246-2.09384 0.4408-0.88162 0.55098-1.54283 0.22034-1.98364-0.22034-0.3306-0.66119-0.77142-1.43263-1.21223-0.66123-0.44081-1.10203-0.66121-1.54284-0.66121-0.55102-0.11021-1.10199 0.22042-1.65301 0.88162-0.22035 0.22042-0.66123 0.99182-1.54284 2.31425 0 0-5.06929 7.93459-5.06929 7.93459l4.62848 2.75504m31.73828-37.6892-4.18772 0.33059-5.51009 13.44469s2.75505 3.19586 2.75505 3.19586l-2.86526 7.05295-15.97932-19.065s3.96726-9.6978 3.96726-9.6978l24.79551-2.31425s-2.97543 7.05295-2.97543 7.05295m-9.47743 0.55101-9.36718 0.77141-0.55102 1.21223s6.17133 7.05295 6.17133 7.05295l3.74687-9.03659m13.2243-10.79983s-21.15885-6.17133-21.15885-6.17133l4.4081-15.20792c0.99182-3.19587 1.54284-5.17951 1.76323-5.73052 0.4408-1.32243 0.99182-2.31425 1.65305-2.86526 0.66119-0.55102 1.5428-0.88162 2.42445-0.88162 0.88161 0 1.76323 0.22042 2.75505 0.44081 1.21221 0.33059 2.09382 0.77142 2.75505 1.21223 0.88161 0.66121 1.32242 1.76323 1.32242 3.08566 0.4408-0.88162 0.88161-1.54284 1.32245-1.87344 0.99179-0.88162 2.64484-0.88162 4.84887-0.22039 1.76327 0.55101 2.97547 1.10203 3.6367 1.87344 0.55102 0.66122 0.88162 1.54284 0.88162 2.64486 0 0.88162-0.33071 2.31425-0.88162 4.29789l-1.21224 4.18769-4.51828 15.20792m-11.02023-9.91821s2.64484-9.03659 2.64484-9.03659c0.55102-1.98364 0.88161-3.19587 0.77144-3.63668 0-0.55101-0.44081-0.99182-1.32246-1.21222-0.88161-0.22043-1.43263-0.11021-1.76322 0.44081-0.11037 0.22038-0.55102 1.32242-1.10204 3.41627 0 0-2.64484 9.03659-2.64484 9.03659l3.41628 0.99182m7.71417 2.20405 2.64483-9.1468c0.44081-1.65303 0.77144-2.64485 0.88162-2.97546 0.22034-0.99182 0.22034-1.65303 0-1.98364-0.22035-0.22043-0.55098-0.44081-1.102-0.66122-0.88165-0.22042-1.54284-0.11021-1.87344 0.55102-0.11036 0.22042-0.55101 1.54283-1.32245 4.07748l-2.64484 9.14679s3.41628 0.99183 3.41628 0.99183m-9.69782-27.44038s1.21221-6.61214 1.21221-6.61214l9.91824 1.87344c2.31424 0.44081 3.74687 0.55101 4.4081 0.55101 1.10199 0 1.76322-0.66122 2.20403-1.76324 0.33071-1.10202 0.7714-2.75506 1.21221-5.06931 0.4408-2.31425 0.66123-3.96728 0.77144-4.9591 0-1.43263-0.55102-2.31425-1.98365-2.75506-1.10204-0.33059-2.31424-0.66121-3.74687-0.88161l-9.91824-1.87344s1.21225-6.61214 1.21225-6.61214l9.91819 1.87344c3.41628 0.77141 5.73052 1.32242 6.72235 1.65303 1.76322 0.55101 2.86526 1.76324 3.52649 3.30607 0.4408 1.21223 0.4408 3.08567 0.11036 5.73052-0.11036 1.21222-0.44085 3.08566-0.99183 5.62032-0.66122 3.74688-1.32245 6.28153-1.87343 7.82436-0.33071 1.10203-0.77144 2.09385-1.21225 2.64486-0.4408 0.66121-0.88161 1.21222-1.54284 1.54283-0.88161 0.55101-1.76322 0.77142-2.86526 0.77142-1.21221 0-2.75505-0.22042-4.51827-0.55101 0 0-2.75509-0.44081-2.75509-0.44081l-9.80799-1.87344m5.73052-28.76281-0.11036-11.79165 16.30995-11.79165v-0.66121l-16.42016 0.3306s-0.11036-6.83255-0.11036-6.83255l22.04046-0.22038 0.11037 11.68144-16.30995 11.68145v0.77141l16.42016-0.33059 0.11036 6.83254s-22.04047 0.3306-22.04047 0.3306m-0.7714-32.50969-2.86526-26.55875s5.62031-0.66122 5.62031-0.66122l1.10204 9.69781 16.3099-1.76324 0.77144 7.05295-16.30994 1.76324 1.10203 9.808s-5.73052 0.66121-5.73052 0.66121m-2.64487-26.66896s-1.76323-6.50193-1.76323-6.50193l9.80799-2.64486c2.20407-0.66121 3.6367-1.10202 4.18772-1.43263 0.99182-0.44081 1.32241-1.32243 1.2122-2.64486-0.11036-1.10202-0.55101-2.86526-1.10203-5.0693-0.66119-2.31425-1.102-3.85708-1.5428-4.84891-0.55102-1.32242-1.54285-1.87343-2.97548-1.65303-1.10203 0.22042-2.42445 0.44081-3.74687 0.88162l-9.80802 2.64485s-1.76323-6.50193-1.76323-6.50193l9.80799-2.64486c3.41628-0.88162 5.73052-1.32243 6.72234-1.54283 1.87344-0.22042 3.41628 0.22042 4.62853 1.43263 0.88161 0.88162 1.76322 2.64485 2.64483 5.06931 0.33071 1.10202 0.88162 2.97546 1.54284 5.39991 0.99183 3.63668 1.65302 6.28153 1.76323 7.82436 0.11036 1.21223 0.22035 2.20405 0.11036 2.97547-0.11036 0.77141-0.33071 1.43263-0.66123 2.09384-0.55101 0.88162-1.32242 1.54283-2.31424 1.98364-1.10203 0.55102-2.53466 0.99182-4.18768 1.54284 0 0-2.75505 0.77141-2.75505 0.77141l-9.80802 2.86526m11.90184-38.35041s-20.2772 8.70599-20.2772 8.70599l-6.39175-14.87732c-1.54284-3.52647-2.31424-5.62032-2.53466-6.28153-0.22035-0.77142-0.33071-1.54283-0.22035-2.20404 0.11036-0.77142 0.33071-1.43263 0.7714-1.98364 0.44081-0.55102 0.99183-1.10203 1.54284-1.43263 0.66123-0.44081 1.54285-0.88162 2.64484-1.32243 1.65305-0.66122 2.86526-1.10203 3.6367-1.10203 1.102-0.11021 2.09382 0.3306 2.86526 1.10203 0.22035 0.22042 0.55102 0.66121 0.99182 1.32243-0.4408-1.32243-0.4408-2.42446-0.11036-3.30607 0.22035-0.3306 0.44081-0.66122 0.7714-0.99183 0.33071-0.22042 0.88165-0.55101 1.54284-0.88161 0.33071-0.22043 1.10204-0.44081 2.20407-0.99183l1.87344-0.77141s2.64484 6.06113 2.64484 6.06113l-1.10204 0.44081c-1.2122 0.55101-2.09382 0.99182-2.53462 1.32242-0.66123 0.44081-0.99183 1.10203-0.77144 1.76324 0.11036 0.44081 0.4408 1.43263 0.99182 2.75506 0 0 3.74687 8.70598 3.74687 8.70598l5.17954-2.20405s2.53463 6.17133 2.53463 6.17133m-12.89367-1.76323s-3.74687-8.70598-3.74687-8.70598c-0.55102-1.32244-0.88161-2.09385-0.99183-2.31426-0.4408-0.88162-0.88161-1.32242-1.54284-1.32242-0.4408 0-0.99182 0.11021-1.76322 0.4408-0.66123 0.3306-1.21221 0.55102-1.43263 0.88162-0.44081 0.44082-0.44081 0.99182-0.22035 1.87344 0.11036 0.3306 0.44085 1.10203 0.99183 2.53466 0 0 3.74691 8.70598 3.74691 8.70598l4.95908-2.09384'
-
-
-
-
-
-function BrandSvg (props) {
-  return <Svg {...props} id="brandSvg" path={path} viewBox="0 0 512 512" />
-}
-
-
-
-
-
-export default BrandSvg
