--- src/components/svg/CodeRedSvg.js
@@ -1,29 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const path = 'M107.61,92.257c-36.687,0 -66.36,29.673 -66.36,66.36c0,36.688 29.673,66.361 66.36,66.361c36.687,0 66.361,-29.673 66.361,-66.361c0.27,-36.417 -29.674,-66.36 -66.361,-66.36Zm0,21.311c25.088,0 45.32,20.232 45.32,45.319c0,25.088 -20.232,45.32 -45.32,45.32c-25.087,0 -45.319,-20.232 -45.319,-45.32c0,-25.087 20.232,-45.319 45.319,-45.319Zm70.407,-80.927c-20.232,0 -36.687,16.455 -36.687,36.687c0,20.232 16.455,36.687 36.687,36.687c20.232,0 36.687,-16.455 36.687,-36.687c0,-20.232 -16.455,-36.687 -36.687,-36.687Zm0,11.599c13.758,0 24.818,11.06 24.818,24.818c0,13.758 -11.06,24.818 -24.818,24.818c-13.758,0 -24.818,-11.06 -24.818,-24.818c-0.269,-13.488 11.06,-24.818 24.818,-24.818Zm9.981,206.095l0,5.665l-31.831,0c0,-1.349 0.27,-2.698 0.809,-4.046c0.809,-2.158 2.158,-4.316 3.777,-6.475c1.888,-2.158 4.316,-4.585 7.823,-7.283c5.125,-4.316 8.902,-7.823 10.52,-10.251c1.619,-2.428 2.698,-4.855 2.698,-7.013c0,-2.428 -0.81,-4.317 -2.428,-5.935c-1.619,-1.619 -3.777,-2.428 -6.474,-2.428c-2.968,0 -5.126,0.809 -6.744,2.698c-1.619,1.888 -2.698,4.046 -2.698,7.013l-6.204,-0.539c0.539,-4.586 1.888,-8.093 4.586,-10.521c2.697,-2.428 6.474,-3.507 11.06,-3.507c4.586,0 8.362,1.349 11.06,3.777c2.697,2.698 4.046,5.665 4.046,9.442c0,1.888 -0.27,3.776 -1.079,5.664c-0.809,1.889 -2.158,3.777 -4.046,5.935c-1.889,2.158 -4.856,4.856 -9.172,8.363c-3.507,2.967 -5.935,5.125 -6.744,6.204c-1.079,1.079 -1.888,2.158 -2.428,3.237l23.469,0Zm-49.905,-250.335c-10.79,0 -19.692,8.902 -19.692,19.692c0,10.791 8.902,19.693 19.692,19.693c10.79,0 19.692,-8.902 19.692,-19.693c0,-11.06 -8.902,-19.692 -19.692,-19.692Zm0,6.204c7.553,0 13.488,5.935 13.488,13.488c0,7.554 -5.935,13.488 -13.488,13.488c-7.553,0 -13.488,-5.934 -13.488,-13.488c0,-7.553 5.935,-13.488 13.488,-13.488Z'
-
-
-
-
-function CodeRedSvg (props) {
-  return <Svg {...props} path={path} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default CodeRedSvg
