--- src/components/svg/DispatchSvg.js
@@ -1,30 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const path = 'M142.871,193.513l-53.806,-82.501c0,0 13.245,-13.521 43.045,-25.661c29.8,-12.141 41.94,-12.141 41.94,-12.141l20.971,95.746l60.979,-0.828c0,0 -36.146,7.45 -77.259,28.972c-40.837,21.522 -74.776,48.563 -74.776,48.563l38.906,-52.15Zm-43.321,-35.87c1.932,-1.38 2.484,-3.863 1.104,-5.794l-30.076,-46.08c3.587,-3.035 11.037,-9.106 22.902,-16.28c17.383,-10.761 34.491,-16.279 34.491,-16.279c2.207,-0.828 3.311,-3.035 2.759,-5.243c-0.828,-2.207 -3.035,-3.311 -5.243,-2.759c-0.827,0.276 -18.211,5.794 -36.422,17.107c-18.211,11.313 -26.489,19.591 -26.764,19.867c-1.38,1.38 -1.656,3.587 -0.552,5.243l32.007,48.838c0.828,1.104 2.207,1.932 3.587,1.932c0.552,0 1.38,-0.276 2.207,-0.552Zm-10.761,22.35c1.932,-1.38 2.484,-3.863 1.104,-5.795l-49.391,-73.672c5.519,-4.966 19.591,-16.555 39.182,-28.144c25.385,-15.176 52.15,-24.281 52.426,-24.281c2.207,-0.828 3.311,-3.036 2.759,-5.243c-0.828,-2.207 -3.035,-3.311 -5.243,-2.759c-1.103,0.276 -27.868,9.381 -54.081,25.109c-25.937,15.728 -42.493,30.904 -43.044,31.455c-1.38,1.38 -1.656,3.588 -0.552,5.243l51.322,76.983c0.828,1.104 2.207,1.932 3.311,1.932c0.552,0 1.38,-0.276 2.207,-0.828Zm-3.587,33.111c1.932,-1.38 2.484,-3.863 1.104,-5.795l-76.707,-112.853c7.726,-6.898 31.731,-26.765 60.152,-43.044c34.49,-19.591 74.223,-33.111 74.775,-33.111c2.208,-0.828 3.311,-3.035 2.484,-5.243c-0.828,-2.207 -3.036,-3.311 -5.243,-2.483c-0.276,0 -40.837,13.796 -76.155,33.939c-35.319,20.142 -64.015,46.079 -64.291,46.355c-1.655,1.38 -1.655,3.587 -0.552,5.519l78.915,115.888c0.828,1.104 2.207,1.932 3.311,1.932c0.828,-0.276 1.655,-0.552 2.207,-1.104Z'
-
-
-
-
-
-function DispatchSvg (props) {
-  return <Svg {...props} path={path} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default DispatchSvg
