--- src/components/svg/FirstYearSvg.js
@@ -1,28 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const path = 'M15.071,127.289l-13.679,3.1c-0.365,-1.641 -0.729,-2.735 -0.729,-3.1c-0.912,-4.377 -0.912,-7.478 0.182,-9.666c0.912,-2.007 2.736,-3.648 5.472,-4.742c3.465,-1.642 10.578,-3.83 21.338,-6.201c7.296,-1.642 12.22,-2.554 15.138,-2.736c3.283,-0.365 5.837,0 7.66,0.912c1.824,0.912 3.101,2.188 3.831,3.83c0.911,1.641 1.459,3.465 2.006,5.654c0.729,3.1 0.729,5.471 0.182,7.113c-0.729,2.006 -2.553,3.647 -5.654,5.106c2.918,0 4.925,0.183 6.019,0.73c1.277,0.547 2.371,1.641 3.283,3.1c0.729,1.277 1.276,3.283 2.006,5.837c0.912,3.647 0.912,6.566 0,8.572c-0.912,2.006 -2.553,3.647 -5.107,5.107c-3.465,1.823 -12.037,4.377 -25.898,7.295c-5.472,1.277 -9.667,1.824 -12.22,2.006c-3.648,0.182 -6.383,-0.365 -8.207,-1.277c-1.824,-1.094 -3.101,-2.553 -4.013,-4.924c-0.729,-1.824 -1.459,-4.742 -2.371,-8.39l13.679,-3.1c0.183,0.365 0.183,0.729 0.183,0.912c0.547,2.006 0.911,3.283 1.459,3.83c0.364,0.547 0.912,0.729 1.641,0.912c0.73,0 1.642,0 3.283,-0.365c1.094,-0.182 4.195,-0.912 9.302,-2.006c6.018,-1.459 9.484,-2.371 10.213,-2.736c2.006,-0.547 3.101,-1.459 3.465,-2.371c0.183,-0.547 0.183,-1.459 0,-2.553c-0.364,-2.007 -1.459,-2.919 -2.918,-3.101c-1.094,0 -4.559,0.547 -10.396,1.824l-8.389,1.824l-2.371,-10.214l10.943,-2.371c3.83,-0.912 6.383,-1.641 7.113,-2.553c0.912,-0.73 1.094,-2.006 0.729,-3.465c-0.365,-1.824 -1.276,-2.736 -2.553,-2.919c-1.277,-0.182 -4.925,0.365 -10.761,1.824c-7.295,1.642 -11.49,2.736 -12.949,3.283c-1.277,0.547 -2.006,1.459 -1.824,2.736c0.182,0 0.547,1.277 0.912,3.283Zm70.035,-11.49l-14.043,0.912c-0.183,-1.642 -0.183,-2.736 -0.183,-3.101c-0.182,-4.377 0.365,-7.66 1.824,-9.484c1.277,-1.824 3.283,-3.1 6.019,-3.83c3.648,-1.094 11.125,-2.006 22.068,-2.736c7.296,-0.364 12.585,-0.547 15.503,-0.364c3.283,0.182 5.654,0.912 7.295,2.188c1.642,1.094 2.736,2.554 3.283,4.377c0.547,1.824 0.912,3.648 1.095,5.837c0.182,3.1 -0.183,5.471 -0.912,6.93c-0.912,1.824 -3.101,3.283 -6.384,4.195c2.918,0.547 4.742,1.094 5.837,1.642c1.276,0.729 2.188,2.006 2.735,3.465c0.547,1.459 0.73,3.465 0.912,6.019c0.183,3.83 -0.182,6.565 -1.459,8.572c-1.277,1.823 -3.1,3.282 -5.836,4.194c-3.648,1.277 -12.585,2.371 -26.628,3.283c-5.654,0.365 -9.849,0.365 -12.402,0c-3.648,-0.364 -6.201,-1.276 -7.843,-2.553c-1.641,-1.277 -2.553,-3.101 -3.1,-5.472c-0.365,-2.006 -0.73,-4.924 -0.912,-8.754l14.043,-0.912c0,0.365 0,0.73 0,0.912c0.183,2.006 0.365,3.465 0.73,4.012c0.365,0.548 0.912,0.912 1.459,1.095c0.547,0.182 1.641,0.182 3.283,0.182c1.094,0 4.377,-0.182 9.666,-0.365c6.201,-0.547 9.849,-0.912 10.578,-1.094c2.007,-0.365 3.283,-0.912 3.83,-1.824c0.365,-0.547 0.365,-1.459 0.365,-2.553c-0.182,-2.006 -0.912,-3.101 -2.371,-3.465c-1.094,-0.183 -4.742,-0.183 -10.578,0.182l-8.572,0.547l-0.73,-10.396l11.126,-0.729c4.012,-0.183 6.566,-0.73 7.478,-1.459c0.911,-0.73 1.459,-1.824 1.276,-3.283c-0.182,-1.824 -0.729,-2.918 -2.006,-3.283c-1.277,-0.365 -4.924,-0.365 -10.943,0c-7.478,0.365 -11.855,0.912 -13.314,1.276c-1.459,0.365 -2.189,1.095 -2.189,2.554c0,-0.183 0,1.094 0,3.283Zm86.268,-18.603c9.119,0.547 15.685,1.094 19.515,2.006c4.377,0.912 7.295,2.918 8.754,6.201c0.912,2.006 1.277,4.742 1.277,8.572c0,1.094 -0.182,4.377 -0.547,9.849c-0.365,6.565 -0.73,10.76 -0.912,12.219c-0.547,3.648 -1.824,6.384 -3.648,8.025c-2.006,1.824 -4.559,2.736 -7.66,3.101c-3.83,0.365 -10.396,0.182 -19.697,-0.183c-9.302,-0.364 -15.685,-1.276 -19.515,-2.006c-4.377,-0.912 -7.296,-2.918 -8.755,-6.201c-0.912,-2.006 -1.276,-4.924 -1.276,-8.572c0,-1.094 0.182,-4.377 0.547,-9.666c0.365,-6.748 0.729,-10.943 0.912,-12.402c0.547,-3.648 1.824,-6.384 3.647,-8.025c2.007,-1.824 4.56,-2.736 7.66,-3.101c3.831,-0.365 10.396,-0.365 19.698,0.183Zm-1.642,12.766c-6.748,-0.364 -10.943,-0.182 -12.402,0.548c-1.094,0.547 -1.823,1.823 -2.371,3.647c-0.364,1.459 -0.547,3.83 -0.729,6.931c-0.183,4.195 -0.183,7.113 0.365,8.754c0.364,1.459 1.276,2.371 2.735,2.918c1.095,0.548 2.919,0.73 5.29,0.912l5.653,0.547c6.019,0.183 9.484,0.365 10.396,0.365c1.642,-0.182 2.918,-0.365 3.648,-0.729c0.912,-0.365 1.459,-1.095 1.824,-2.189c0.547,-1.459 1.094,-4.377 1.276,-8.572c0.183,-4.195 0.183,-7.113 -0.364,-8.572c-0.365,-1.459 -1.277,-2.371 -2.736,-2.918c-1.277,-0.912 -5.654,-1.459 -12.585,-1.642Zm63.288,-5.106l22.98,4.559l-9.484,47.785l-14.591,-2.918l6.931,-35.2l-3.83,-0.73l-16.415,9.119l-6.201,-10.943l20.61,-11.672Z'
-
-
-
-function FirstYearSvg (props) {
-  return <Svg {...props} path={path} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default FirstYearSvg
