--- src/components/svg/Rescue1000Svg.js
@@ -1,33 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const paths = [
-  'M197.862,227.326c0,0 -26.522,-11.943 -69.203,-12.207c-42.682,-0.263 -72.278,12.207 -72.278,12.207l-5.005,-26.346l10.187,-13.261l14.315,10.89l13.085,-15.984l16.511,10.89l23.887,-24.063l24.327,24.326l15.193,-11.504l11.768,14.051l14.754,-9.397l7.026,13.964l-4.567,26.434Z',
-  'M197.511,90.412c0.79,2.723 1.141,5.621 1.141,8.519c0,16.159 -12.207,29.245 -27.4,29.245c-15.193,0 -27.4,-13.086 -27.4,-29.245c0,-14.315 9.66,-26.171 22.394,-28.718l-38.29,-41.539l-38.29,41.539c12.734,2.459 22.394,14.403 22.394,28.718c0,16.159 -12.207,29.245 -27.4,29.245c-15.193,0 -27.4,-13.086 -27.4,-29.245c0,-2.898 0.439,-5.796 1.141,-8.519l-58.401,-22.833l47.775,113.641l11.505,-14.754l15.72,12.119l13.173,-18.091l15.808,12.734l24.59,-28.366l25.292,26.873l16.774,-10.978l12.383,16.95l13.613,-10.363l10.099,13.525l49.268,-113.29l-58.489,22.833Z',
-]
-
-
-
-
-
-function Rescue1000Svg (props) {
-  return <Svg {...props} path={paths} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default Rescue1000Svg
