--- src/components/svg/Rescue100Svg.js
@@ -1,30 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const path = 'M127.068,0c-61.477,58.724 -85.64,106.743 -85.64,106.743c-7.646,15.293 -11.928,32.421 -11.928,49.549c0,55.053 44.043,99.708 98.485,99.708c54.442,0 98.485,-44.655 98.485,-99.708c0,-17.74 -5.811,-34.562 -12.54,-50.466c-15.293,-36.397 -86.862,-105.826 -86.862,-105.826Zm0,101.849c8.869,0 15.904,7.341 15.904,16.211l0,17.128l16.822,0c8.87,0 15.904,7.34 15.904,16.21c0,8.87 -7.034,16.21 -15.904,16.21l-16.822,0l0,17.128c0,8.87 -7.035,16.21 -15.904,16.21c-8.87,0 -15.905,-7.34 -15.905,-16.21l0,-17.128l-16.822,0c-8.87,0 -15.904,-7.34 -15.904,-16.21c0,-8.87 7.034,-16.21 15.904,-16.21l16.822,0l0,-17.128c0,-8.87 7.035,-16.211 15.905,-16.211Z'
-
-
-
-
-
-function Rescue100Svg (props) {
-  return <Svg {...props} path={path} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default Rescue100Svg
