--- src/components/svg/Rescue200Svg.js
@@ -1,33 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const paths = [
-  'M63.404,0c-39.951,38.161 -55.652,69.366 -55.652,69.366c-4.969,9.938 -7.752,21.069 -7.752,32.199c0,35.777 28.621,64.795 64,64.795c35.379,0 64,-29.018 64,-64.795c0,-11.528 -3.776,-22.459 -8.149,-32.795c-9.938,-23.652 -56.447,-68.77 -56.447,-68.77Zm0,66.186c5.764,0 10.335,4.771 10.335,10.534l0,11.131l10.932,0c5.764,0 10.335,4.77 10.335,10.534c0,5.764 -4.571,10.534 -10.335,10.534l-10.932,0l0,11.131c0,5.764 -4.571,10.534 -10.335,10.534c-5.764,0 -10.336,-4.77 -10.336,-10.534l0,-11.131l-10.931,0c-5.764,0 -10.336,-4.77 -10.336,-10.534c0,-5.764 4.572,-10.534 10.336,-10.534l10.931,0l0,-11.131c0,-5.763 4.572,-10.534 10.336,-10.534Z',
-  'M191.404,89.64c-39.951,38.161 -55.652,69.366 -55.652,69.366c-4.969,9.938 -7.752,21.069 -7.752,32.199c0,35.776 28.621,64.795 64,64.795c35.379,0 64,-29.019 64,-64.795c0,-11.528 -3.776,-22.46 -8.149,-32.795c-9.938,-23.652 -56.447,-68.77 -56.447,-68.77Zm0,66.186c5.764,0 10.335,4.77 10.335,10.534l0,11.131l10.932,0c5.764,0 10.335,4.77 10.335,10.534c0,5.764 -4.571,10.534 -10.335,10.534l-10.932,0l0,11.13c0,5.764 -4.571,10.535 -10.335,10.535c-5.764,0 -10.336,-4.771 -10.336,-10.535l0,-11.13l-10.931,0c-5.764,0 -10.336,-4.77 -10.336,-10.534c0,-5.764 4.572,-10.534 10.336,-10.534l10.931,0l0,-11.131c0,-5.764 4.572,-10.534 10.336,-10.534Z',
-]
-
-
-
-
-
-function Rescue200Svg (props) {
-  return <Svg {...props} path={paths} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default Rescue200Svg
