--- src/components/svg/Rescue300Svg.js
@@ -1,34 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const paths = [
-  'M127.488,0c-34.333,32.795 -47.827,59.612 -47.827,59.612c-4.27,8.54 -6.661,18.105 -6.661,27.671c0,30.745 24.596,55.683 55,55.683c30.404,0 55,-24.938 55,-55.683c0,-9.907 -3.245,-19.302 -7.003,-28.184c-8.54,-20.326 -48.509,-59.099 -48.509,-59.099Zm0,56.879c4.953,0 8.882,4.099 8.882,9.053l0,9.565l9.394,0c4.953,0 8.882,4.099 8.882,9.053c0,4.953 -3.929,9.052 -8.882,9.052l-9.394,0l0,9.566c0,4.953 -3.929,9.052 -8.882,9.052c-4.954,0 -8.882,-4.099 -8.882,-9.052l0,-9.566l-9.395,0c-4.953,0 -8.882,-4.099 -8.882,-9.052c0,-4.954 3.929,-9.053 8.882,-9.053l9.395,0l0,-9.565c0,-4.954 3.928,-9.053 8.882,-9.053Z',
-  'M200.488,113.034c-34.333,32.795 -47.827,59.612 -47.827,59.612c-4.27,8.54 -6.661,18.106 -6.661,27.671c0,30.745 24.596,55.683 55,55.683c30.404,0 55,-24.938 55,-55.683c0,-9.907 -3.245,-19.301 -7.003,-28.183c-8.54,-20.327 -48.509,-59.1 -48.509,-59.1Zm0,56.879c4.953,0 8.882,4.099 8.882,9.053l0,9.565l9.394,0c4.953,0 8.882,4.099 8.882,9.053c0,4.953 -3.929,9.053 -8.882,9.053l-9.394,0l0,9.565c0,4.953 -3.929,9.053 -8.882,9.053c-4.954,0 -8.882,-4.1 -8.882,-9.053l0,-9.565l-9.395,0c-4.953,0 -8.882,-4.1 -8.882,-9.053c0,-4.954 3.929,-9.053 8.882,-9.053l9.395,0l0,-9.565c0,-4.954 3.928,-9.053 8.882,-9.053Z',
-  'M54.488,113.034c-34.333,32.795 -47.827,59.612 -47.827,59.612c-4.27,8.54 -6.661,18.106 -6.661,27.671c0,30.745 24.596,55.683 55,55.683c30.404,0 55,-24.938 55,-55.683c0,-9.907 -3.245,-19.301 -7.003,-28.183c-8.54,-20.327 -48.509,-59.1 -48.509,-59.1Zm0,56.879c4.953,0 8.882,4.099 8.882,9.053l0,9.565l9.394,0c4.953,0 8.882,4.099 8.882,9.053c0,4.953 -3.929,9.053 -8.882,9.053l-9.394,0l0,9.565c0,4.953 -3.929,9.053 -8.882,9.053c-4.954,0 -8.882,-4.1 -8.882,-9.053l0,-9.565l-9.395,0c-4.953,0 -8.882,-4.1 -8.882,-9.053c0,-4.954 3.929,-9.053 8.882,-9.053l9.395,0l0,-9.565c0,-4.954 3.928,-9.053 8.882,-9.053Z',
-]
-
-
-
-
-
-function Rescue300Svg (props) {
-  return <Svg {...props} path={paths} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default Rescue300Svg
