--- src/components/svg/Rescue400Svg.js
@@ -1,35 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const paths = [
-  'M127.581,0c-28.09,26.832 -39.131,48.773 -39.131,48.773c-3.493,6.988 -5.45,14.814 -5.45,22.64c0,25.155 20.124,45.559 45,45.559c24.876,0 45,-20.404 45,-45.559c0,-8.106 -2.655,-15.792 -5.73,-23.059c-6.987,-16.63 -39.689,-48.354 -39.689,-48.354Zm0,46.537c4.053,0 7.267,3.354 7.267,7.407l0,7.826l7.686,0c4.053,0 7.267,3.354 7.267,7.407c0,4.053 -3.214,7.407 -7.267,7.407l-7.686,0l0,7.826c0,4.053 -3.214,7.407 -7.267,7.407c-4.053,0 -7.267,-3.354 -7.267,-7.407l0,-7.826l-7.687,0c-4.052,0 -7.267,-3.354 -7.267,-7.407c0,-4.053 3.215,-7.407 7.267,-7.407l7.687,0l0,-7.826c0,-4.053 3.214,-7.407 7.267,-7.407Z',
-  'M197.581,69.514c-28.09,26.832 -39.131,48.773 -39.131,48.773c-3.493,6.988 -5.45,14.814 -5.45,22.64c0,25.155 20.124,45.559 45,45.559c24.876,0 45,-20.404 45,-45.559c0,-8.106 -2.655,-15.792 -5.73,-23.059c-6.987,-16.63 -39.689,-48.354 -39.689,-48.354Zm0,46.537c4.053,0 7.267,3.354 7.267,7.407l0,7.826l7.686,0c4.053,0 7.267,3.354 7.267,7.407c0,4.053 -3.214,7.407 -7.267,7.407l-7.686,0l0,7.826c0,4.053 -3.214,7.407 -7.267,7.407c-4.053,0 -7.267,-3.354 -7.267,-7.407l0,-7.826l-7.687,0c-4.052,0 -7.267,-3.354 -7.267,-7.407c0,-4.053 3.215,-7.407 7.267,-7.407l7.687,0l0,-7.826c0,-4.053 3.214,-7.407 7.267,-7.407Z',
-  'M57.581,69.514c-28.09,26.832 -39.131,48.773 -39.131,48.773c-3.493,6.988 -5.45,14.814 -5.45,22.64c0,25.155 20.124,45.559 45,45.559c24.876,0 45,-20.404 45,-45.559c0,-8.106 -2.655,-15.792 -5.73,-23.059c-6.987,-16.63 -39.689,-48.354 -39.689,-48.354Zm0,46.537c4.053,0 7.267,3.354 7.267,7.407l0,7.826l7.686,0c4.053,0 7.267,3.354 7.267,7.407c0,4.053 -3.214,7.407 -7.267,7.407l-7.686,0l0,7.826c0,4.053 -3.214,7.407 -7.267,7.407c-4.053,0 -7.267,-3.354 -7.267,-7.407l0,-7.826l-7.687,0c-4.052,0 -7.267,-3.354 -7.267,-7.407c0,-4.053 3.215,-7.407 7.267,-7.407l7.687,0l0,-7.826c0,-4.053 3.214,-7.407 7.267,-7.407Z',
-  'M127.581,139.028c-28.09,26.832 -39.131,48.773 -39.131,48.773c-3.493,6.988 -5.45,14.814 -5.45,22.64c0,25.155 20.124,45.559 45,45.559c24.876,0 45,-20.404 45,-45.559c0,-8.106 -2.655,-15.792 -5.73,-23.059c-6.987,-16.63 -39.689,-48.354 -39.689,-48.354Zm0,46.537c4.053,0 7.267,3.354 7.267,7.407l0,7.826l7.686,0c4.053,0 7.267,3.354 7.267,7.407c0,4.053 -3.214,7.407 -7.267,7.407l-7.686,0l0,7.826c0,4.053 -3.214,7.407 -7.267,7.407c-4.053,0 -7.267,-3.354 -7.267,-7.407l0,-7.826l-7.687,0c-4.052,0 -7.267,-3.354 -7.267,-7.407c0,-4.053 3.215,-7.407 7.267,-7.407l7.687,0l0,-7.826c0,-4.053 3.214,-7.407 7.267,-7.407Z',
-]
-
-
-
-
-
-function Rescue400Svg (props) {
-  return <Svg {...props} path={paths} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default Rescue400Svg
