--- src/components/svg/Rescue500Svg.js
@@ -1,30 +1,0 @@
-// Module imports
-import React from 'react'
-
-
-
-
-
-// Component imports
-import Svg from './Svg'
-
-
-
-
-/* eslint-disable max-len */// Yeah there's no getting away from this one
-// Component constants
-const path = 'M248.26,177.042c0,0 -44.948,-20.242 -117.283,-20.689c-72.335,-0.446 -122.493,20.689 -122.493,20.689l-8.484,-44.651l17.265,-22.475l24.261,18.456l22.176,-27.088l27.982,18.456l40.483,-40.782l41.228,41.228l25.749,-19.498l19.944,23.814l25.005,-15.925l11.907,23.665l-7.74,44.8Z'
-
-
-
-
-
-function Rescue500Svg (props) {
-  return <Svg {...props} path={path} viewBox="0 0 256 256" />
-}
-
-
-
-
-
-export default Rescue500Svg
